+++ script/burk-a-brac.mjs
@@ -1,8 +1,8 @@
 // Burk-a-Brac | Copyright (C) 2026 Patrick Burk
 // Licensed under the GNU General Public License v3.0 or later.
 
-import { registerMapNotesSettings } from "./map-notes.mjs";
-// FIX import { applyMapNotesTweaks, registerMapNotesSettings } from "./map-notes.mjs";
+import { registerJournalHooks, registerJournalSettings } from "./journal.mjs";
+import { registerMapNoteHooks, registerMapNoteSettings } from "./map-note.mjs";
 
 Hooks.once("init", function() {
     console.log("Burk-a-Brac | Module loading");
@@ -12,7 +12,8 @@ Hooks.once("init", function() {
 });
 
 function registerHooks() {
-    // FIX
+    registerJournalHooks();
+    registerMapNoteHooks();
 }
 
 function registerSettings() {
@@ -28,5 +29,46 @@ function registerSettings() {
     //     scope: "world",
     //     type: Number|String|Boolean|Object|DataModel|DataField
     // });
-    registerMapNotesSettings();
+    registerJournalSettings();
+    registerMapNoteSettings();
+    // game.settings.register("burk-a-brac", "UIDefaultCursor", {
+    //     config: true,
+    //     default: "", // Is this required?
+    //     hint: "BaB.UIDefaultCursor_Hint",
+    //     name: "BaB.UIDefaultCursor",
+    // //     onChange: () => {},
+    // //     // requiresReload: true,
+    // //     scope: "world",
+    //     type: String
+    // });
+    // game.settings.register("burk-a-brac", "UIPointerCursor", {
+    //     config: true,
+    //     default: "",
+    //     hint: "BaB.UIPointerCursor_Hint",
+    //     name: "BaB.UIPointerCursor",
+    // //     onChange: () => {},
+    // //     // requiresReload: true,
+    //     scope: "world",
+    //     type: String
+    // });
+    // game.settings.register("burk-a-brac", "UIGrabCursor", {
+    //     config: true,
+    //     default: "",
+    //     hint: "BaB.UIGrabCursor_Hint",
+    //     name: "BaB.UIGrabCursor",
+    // //     onChange: () => {},
+    // //     // requiresReload: true,
+    //     scope: "world",
+    //     type: String
+    // });
+    // game.settings.register("burk-a-brac", "UITextCursor", {
+    //     config: true,
+    //     default: "",
+    //     hint: "BaB.UITextCursor_Hint",
+    //     name: "BaB.UITextCursor",
+    // //     onChange: () => {},
+    // //     // requiresReload: true,
+    //     scope: "world",
+    //     type: String
+    // });
 }