+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v1.0.1
 
+- Added settings for displaying journal thumbnails in the sidebar (inspired by [Journal Thumbnails](https://foundryvtt.com/packages/journal-thumbnail)).
 - Added badges and setting info to README.
 - Added license info to scripts.
 - Moved settings registration to individual files.